Split out metadata extraction
